# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/semesters": {
            "get": {
                "tags": [
                    "Semester"
                ],
                "summary": "SemestersInfo",
                "operationId": "cf2913192f19f5e219c0b890da6f57af",
                "parameters": [
                    {
                        "name": "group",
                        "in": "query",
                        "description": "Group name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "OAIp2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "semesters": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "number": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "start_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "1994-10-11"
                                                    },
                                                    "end_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2002-01-19"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Semester"
                ],
                "summary": "Create",
                "operationId": "ac66ba2c9916d7ac9fc6cc8ff4d58e60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1994-10-11"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2002-01-19"
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "semester": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "number": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1994-10-11"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2002-01-19"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/semesters/{semester}": {
            "put": {
                "tags": [
                    "Semester"
                ],
                "summary": "Update",
                "operationId": "8e4e8ab19b8d2f8f7e242e0cd01a7bc1",
                "parameters": [
                    {
                        "name": "semester",
                        "in": "path",
                        "description": "Semester ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1994-10-11"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2002-01-19"
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "semester": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "number": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1994-10-11"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2002-01-19"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Semester"
                ],
                "summary": "Delete",
                "operationId": "545345072f4df4f5c09c5c75b907e5e3",
                "parameters": [
                    {
                        "name": "semester",
                        "in": "path",
                        "description": "Semester ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}